+++ internal/providers/diagnostic_codes.go
@@ -62,6 +62,7 @@ const (
 	CodeEndForInvalid               = "endfor_invalid"
 	CodeUnknownToken                = "unknown_token"
 	CodeUnexpectedToken             = "unexpected_token"
+	CodeUnterminatedString          = "unterminated_string"
 	CodeCommentTermination          = "comment_termination"
 	CodeCommentTextAfterTerminator  = "comment_text_after_terminator"
 	CodeAssignmentInCondition       = "assignment_in_condition"